Apply @ihhub suggestions
